The changes made in this commit include: - Added [file name] - Modified [file name] to [brief description of change] - Deleted [file name]
In the `quickPickSetupAzureOpenai.ts` file, the code was modified to remove the check for `OPENAI_OAUTH2_STARTSWITH` as it is no longer needed. The function now only checks if the length of the `name` parameter is greater than or equal to `OPENAI_APIKEY_LENGTH` to determine if the API key or token is valid.